# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plymotion API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/videos/inspect": {
      "post": {
        "tags": [
          "convert"
        ],
        "summary": "Inspect Video",
        "operationId": "inspect_video_api_videos_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/convert/estimate": {
      "post": {
        "tags": [
          "convert"
        ],
        "summary": "Estimate",
        "operationId": "estimate_api_convert_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estimate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/convert": {
      "post": {
        "tags": [
          "convert"
        ],
        "summary": "Start Convert",
        "operationId": "start_convert_api_convert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library/{slug}": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Get Library Theme",
        "operationId": "get_library_theme_api_library__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryTheme"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "library"
        ],
        "summary": "Delete Library Theme",
        "operationId": "delete_library_theme_api_library__slug__delete",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/{slug}/frames/{index}": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Library Frame",
        "operationId": "library_frame_api_library__slug__frames__index__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/{slug}/install": {
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Install",
        "operationId": "install_api_library__slug__install_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/{slug}/reveal": {
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Reveal Theme",
        "operationId": "reveal_theme_api_library__slug__reveal_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConvertRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "max_width": {
            "type": "integer",
            "maximum": 7680.0,
            "minimum": 16.0,
            "title": "Max Width",
            "default": 320
          },
          "max_height": {
            "type": "integer",
            "maximum": 4320.0,
            "minimum": 16.0,
            "title": "Max Height",
            "default": 240
          },
          "fps": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Fps",
            "default": 30
          },
          "colors": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 2.0,
            "title": "Colors",
            "default": 64
          },
          "trim_start": {
            "type": "number",
            "minimum": 0.0,
            "title": "Trim Start",
            "default": 0
          },
          "trim_duration": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim Duration"
          }
        },
        "type": "object",
        "required": [
          "path",
          "name"
        ],
        "title": "ConvertRequest"
      },
      "Estimate": {
        "properties": {
          "frame_count": {
            "type": "integer",
            "title": "Frame Count"
          },
          "loop_seconds": {
            "type": "number",
            "title": "Loop Seconds"
          },
          "refresh_rate": {
            "type": "integer",
            "title": "Refresh Rate"
          }
        },
        "type": "object",
        "required": [
          "frame_count",
          "loop_seconds",
          "refresh_rate"
        ],
        "title": "Estimate"
      },
      "EstimateRequest": {
        "properties": {
          "duration": {
            "type": "number",
            "minimum": 0.0,
            "title": "Duration"
          },
          "fps": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Fps"
          },
          "trim_start": {
            "type": "number",
            "minimum": 0.0,
            "title": "Trim Start",
            "default": 0
          },
          "trim_duration": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim Duration"
          }
        },
        "type": "object",
        "required": [
          "duration",
          "fps"
        ],
        "title": "EstimateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobInfo"
          }
        },
        "type": "object",
        "required": [
          "job"
        ],
        "title": "JobAccepted"
      },
      "JobDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "type": "string",
            "enum": [
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "result": {
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cancellable": {
            "type": "boolean",
            "title": "Cancellable"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "log_tail": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Log Tail"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "state",
          "progress",
          "stage",
          "error",
          "cancellable",
          "created_at",
          "finished_at",
          "log_tail",
          "logs"
        ],
        "title": "JobDetail"
      },
      "JobInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "type": "string",
            "enum": [
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "result": {
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cancellable": {
            "type": "boolean",
            "title": "Cancellable"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "log_tail": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Log Tail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "state",
          "progress",
          "stage",
          "error",
          "cancellable",
          "created_at",
          "finished_at",
          "log_tail"
        ],
        "title": "JobInfo"
      },
      "LibraryTheme": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "frame_count": {
            "type": "integer",
            "title": "Frame Count"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "fps": {
            "type": "integer",
            "title": "Fps"
          },
          "colors": {
            "type": "integer",
            "title": "Colors"
          },
          "loop_seconds": {
            "type": "number",
            "title": "Loop Seconds"
          },
          "total_bytes": {
            "type": "integer",
            "title": "Total Bytes"
          },
          "source_video": {
            "type": "string",
            "title": "Source Video"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "installed": {
            "type": "boolean",
            "title": "Installed"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "frame_url_template": {
            "type": "string",
            "title": "Frame Url Template"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "frame_count",
          "width",
          "height",
          "fps",
          "colors",
          "loop_seconds",
          "total_bytes",
          "source_video",
          "created_at",
          "installed",
          "thumbnail_url",
          "frame_url_template"
        ],
        "title": "LibraryTheme"
      },
...
      "PathRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "PathRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoInfo": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "codec": {
            "type": "string",
            "title": "Codec"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url"
          }
        },
        "type": "object",
        "required": [
          "path",
          "name",
          "width",
          "height",
          "duration",
          "fps",
          "codec",
          "size_bytes",
          "media_url"
        ],
        "title": "VideoInfo"
      }
    }
  }
}